+++ Modules/Ruleset.ComputerInfo/Public/Test-UNC.ps1
@@ -36,11 +36,8 @@ Test if UNC (Universal Naming Convention) path is valid
 .PARAMETER Name
 Universal Naming Convention path
 
-.PARAMETER Strict
-If specified produces an error instead of warning for invalid UNC
-
 .PARAMETER Quiet
-Suppresses error or warning message, only true or false is returned
+if specified errors are not shown, only true or false is returned.
 
 .EXAMPLE
 PS> Test-UNC \\SERVER\Share
@@ -63,10 +60,10 @@ PS> Test-UNC \SERVER-01\Share\Directory DIR
 False
 
 .INPUTS
-None. You cannot pipe objects to Test-UNC
+[string]
 
 .OUTPUTS
-None. Test-UNC does not generate any output
+[bool]
 
 .NOTES
 A UNC path can be used to access network resources, and MUST be in the format specified by the
@@ -76,84 +73,84 @@ A valid UNC path MUST contain two or more path components.
 "SERVER" is referred to as the "first pathname component", "Share" as the "second pathname component"
 The size and valid characters for a path component are defined by the protocol used to access the
 resource and the type of resource being accessed.
+
+.LINK
+https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Ruleset.ComputerInfo/Help/en-US/Test-UNC.md
+
+.LINK
+https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file
+
+.LINK
+https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-dfsc/149a3039-98ce-491a-9268-2f5ddef08192
 #>
 function Test-UNC
 {
 	[OutputType([bool])]
-	[CmdletBinding(DefaultParameterSetName = "None",
+	[CmdletBinding(
 		HelpURI = "https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Ruleset.ComputerInfo/Help/en-US/Test-UNC.md")]
 	param (
-		[Parameter(Mandatory = $true, Position = 0)]
-		[string] $Name,
+		[Parameter(Mandatory = $true, Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
+		[string[]] $Name,
 
-		[Parameter(ParameterSetName = "Strict")]
-		[switch] $Strict,
-
-		[Parameter(ParameterSetName = "Quiet")]
+		[Parameter()]
 		[switch] $Quiet
 	)
 
-	Write-Debug -Message "[$($MyInvocation.InvocationName)] params($($PSBoundParameters.Values))"
-
-	# https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file
-	# https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-dfsc/149a3039-98ce-491a-9268-2f5ddef08192
-	if ($Name.Length -gt 260)
+	begin
 	{
-		if (!$Quiet)
+		if ($Quiet)
 		{
-			# The maximum length for a path is 260 characters.
-			# NOTE: Windows 10 version 1607 and later versions of Windows require changing a registry
-			# key or using the Group Policy to remove the limit.
-			if ($Strict)
-			{
-				Write-Error -Category SyntaxError -TargetObject $Name -Message "The maximum length for UNC path is 260 characters"
-			}
-			else
-			{
-				Write-Warning -Message "The maximum length for UNC path is 260 characters"
-			}
-		}
-
-		return $false
-	}
-
-	if ($Name -match "^\\\\\.\\")
-	{
-		if (!$Quiet)
-		{
-			# The "\\.\" prefix will access the Win32 device namespace instead of the Win32 file namespace.
-			if ($Strict)
-			{
-				Write-Error -Category SyntaxError -TargetObject $Name -Message "Specified UNC path bellongs to Win32 device namespace: $Name"
-			}
-			else
-			{
-				Write-Warning -Message "Specified UNC path bellongs to Win32 device namespace: $Name"
-			}
-		}
-
-		return $false
-	}
-
-	# TODO: This regex needs to be verified, ex. space and dot might not need to be present
-	# "^\\\\[a-zA-Z0-9\.\-_]{1,}(\\[a-zA-Z0-9\-_\s\.]{1,}){1,}[\$]{0,1}"
-	if ($Name -match "^\\\\[\w\-_]+(\\[\w\-_]+)+[\$]?")
-	{
-		Write-Debug -Message "[$($MyInvocation.InvocationName)] UNC path syntax verification passed: $Name"
-		return $true
-	}
-
-	if (!$Quiet)
-	{
-		if ($Strict)
-		{
-			Write-Error -Category SyntaxError -TargetObject $Name -Message "UNC path syntax verification failed: $Name"
+			$WriteError = "SilentlyContinue"
 		}
 		else
 		{
-			Write-Warning -Message "UNC path syntax verification failed: $Name"
+			$WriteError = $ErrorActionPreference
 		}
 	}
+	process
+	{
+		Write-Debug -Message "[$($MyInvocation.InvocationName)] params($($PSBoundParameters.Values))"
 
-	return $false
+		foreach ($UNC in $Name)
+		{
+			if ($UNC.Length -gt 260)
+			{
+				# The maximum length for a path is 260 characters.
+				# NOTE: Windows 10 version 1607 and later versions of Windows require changing a registry
+				# key or using the Group Policy to remove the limit.
+				Write-Error -Category SyntaxError -TargetObject $UNC -ErrorAction $WriteError `
+					-Message "The maximum length for UNC path is 260 characters"
+				return $false
+			}
+
+			if ($UNC -match "^\\\\\.\\")
+			{
+				# The "\\.\" prefix will access the Win32 device namespace instead of the Win32 file namespace.
+				Write-Error -Category SyntaxError -TargetObject $UNC -ErrorAction $WriteError `
+					-Message "Specified UNC path bellongs to Win32 device namespace: $UNC"
+
+				return $false
+			}
+
+			# TODO: This regex needs to be verified, ex. space and dot might not need to be present
+			# "^\\\\[a-zA-Z0-9\.\-_]{1,}(\\[a-zA-Z0-9\-_\s\.]{1,}){1,}[\$]{0,1}"
+			# [regex] $Regex = "^\\\\[A-Z0-9\-]+(\\[a-zA-Z0-9\-_\s\.]+)+[\$]?"
+			if ($UNC -notmatch "^\\\\[\w\-_]+(\\[\w\-_]+)+[\$]?")
+			{
+				switch ($ErrorLevel)
+				{
+					0
+					{
+						Write-Error -Category SyntaxError -TargetObject $UNC -ErrorAction $WriteError `
+							-Message "UNC path syntax verification failed for: $UNC"
+					}
+					default {}
+				}
+
+				return $false
+			}
+
+			return $true
+		}
+	}
 }