+++ Readme/CHANGELOG.md
@@ -41,8 +41,10 @@ If you would like to see fresh changes done since last release you can do so on 
   - `Resolve-FileSystem` function to resolve path wildcard pattern or relative paths to `System.IO.*Info`
   - `Reset-TestDrive` function to safely clear test drive directory for cleaner and smoother unit testing.
   - `Resolve-Host` function used to resolve host to IP or an IP to host.
-  - `Test-UPN` function used to validate syntax of Universal Principal Name (UPN)
+  - `Test-UPN` function used to validate syntax of UPN (Universal Principal Name)
   - `Convert-FromSDDL` function to resolve SDDL strings into principal, user name, SID and domain
+  - `NetBiosName` function to validate syntax of NETBIOS name
+  - `TEST-UNC` function to validate syntax of UNC (Universal Naming Convention) name
 
 - Tests